Get compiling against SDL2 on Linux
Modified the build scripts to test for and compile against SDL2, not
SDL1.  Also fixed version checking to use -ge and not -gt.

uint# and sint# are not standard.  I replaced them with uint#_t
and int#_t respectively.

Finally I resolved all build errors with changing the library by
porting over to the SDL2 conventions.  Unfortunately the code is
very biased towards soft-rendering, so the current approach is to
use one texture and one software surface.  In future it would be
better to not do all of the 2D transformations in software.

This version runs but for some reason segfaults loading Vera.  I
think it may be to fo with my modifications to the SDL_RWops handlers.